requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3
idna==2.7
mysql-replication==0.18
pycparser==2.18
PyMySQL==0.9.2
six==1.11.0
